Wait for idle to intersect
