+++ packages/next/client/use-intersection.tsx
@@ -1,4 +1,5 @@
 import { useCallback, useEffect, useRef, useState } from 'react'
+import requestIdleCallback from './request-idle-callback'
 
 type UseIntersectionObserverInit = Pick<IntersectionObserverInit, 'rootMargin'>
 type UseIntersection = { disabled?: boolean } & UseIntersectionObserverInit
@@ -37,7 +38,7 @@ export function useIntersection<T extends Element>({
 
   useEffect(() => {
     if (!hasIntersectionObserver) {
-      if (!visible) setVisible(true)
+      if (!visible) requestIdleCallback(() => setVisible(true))
     }
   }, [visible])
